+++ src/componets/Footer/footer.js
@@ -2,8 +2,6 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faFacebook, faInstagram, faDiscord} from '@fortawesome/free-brands-svg-icons';
-import { } from '@fortawesome/free-brands-svg-icons';
-
 import "./footer.scss"
