+++ src/componets/Header/header.scss
@@ -1,24 +1,41 @@
 header{
   background-color: rgba(0, 0, 0, 0.25);
 }
-.header{
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  width: 90%;
-  align-items: center;
-  margin: auto;
+@media (min-width: 340px) {
+  .header{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    width: 90%;
+    align-items: center;
+    margin: auto;
+  }
+  .header h1{
+    margin: 10px;
+    padding-left: 45px;
+    text-align: center;
+  }
+  .header img{
+    font-size: 2.6em;
+    width: 20%;
+    margin:25px;
+  }
 }
-.header h1{
-  margin: 10px;
-  width: 60%;
-  text-align: center;
+@media (min-width: 580px) {
+  .header h1{
+    font-size: 3em;
+  }
+  .header img{
+    width: 15%;
+    margin:25px;
+  }
 }
-.header img{
-  width: 20%;
-  margin:25px;
-}
-header h3{
-  margin: 0;
-  text-align: center;
+@media (min-width: 992px) { 
+  .header img{
+    width: 10%;
+    margin:25px;
+  }
+  .header h1{
+    font-size: 3.8em;
+  }
 }
