+++ src/scss/custom.scss
@@ -30,7 +30,7 @@
   justify-content: center;
 }
 .navbar{
-  background-color: var(--theme-nav-background);
+  // background-color: var(--theme-nav-background);
   flex-wrap: nowrap;
 }
 .navbar .container{
