+++ src/scss/theme.scss
@@ -3,7 +3,7 @@
   --theme-page-background: rgba(48, 16, 98, 0.60);
   --theme-page-text: rgb(17, 17, 17);
   --theme-page-htext: rgb(0, 0, 0);
-  --theme-nav-background:rgba(138, 13, 6, 0.80);
+  --theme-nav-background:rgba(0, 0, 0, 0.80);
   --theme-card-background: rgba(151, 183, 191, 0.70);
   --theme-footer-background: rgba(138, 13, 6, 0.80);
   --theme-panel-background:rgb(158, 221, 240);
